reimplement sort_id with prototype
