+++ lib/entry.js
@@ -2,6 +2,7 @@ module.exports = require('./kernel');
 
 require('./tasks/timing');
 require('./tasks/error');
+require('./tasks/sort-id');
 
 require('./events/click');
 require('./events/change');
