+++ lib/util/batching.js
@@ -158,39 +158,7 @@ exports.startSendingLoop = function(configs) {
   window.addEventListener('beforeunload', handlePageClose);
 };
 
-// These lines of code are for validating purpose
-// sort_id is mainly used for validating
-var stateSortId = 0;
-var trackSortId = function() {
-  try {
-    var raw = localStorage.getItem(configs.sortIdKey) || '0';
-    stateSortId = parseInt(raw, 10);
-  } catch (error) {
-    // localStorage might fail
-  }
-  window.addEventListener('beforeunload', function() {
-    try {
-      localStorage.setItem(configs.sortIdKey, stateSortId.toString());
-    } catch (error) {
-      // do nothing
-    }
-  });
-};
-var loopMaxNumber = 1000000;
-var incSortId = function() {
-  if (stateSortId > loopMaxNumber) {
-    stateSortId = 0;
-  } else {
-    stateSortId += 1;
-  }
-}
-
 exports.trackEvent = function(event) {
-  incSortId();
-  event.sort_id = stateSortId;
   stateEventQueue.push(event);
   remindSendingMessage();
 };
-
-// set sort_id for checking purpose
-trackSortId();
